Add button component initial version
