+++ src/components/Button/interface.tsx
@@ -1,0 +1,3 @@
+export interface ButtonProps {
+  primary: boolean;
+}
