+++ src/components/Wrapper/interface.ts
@@ -1,0 +1,12 @@
+interface WraperProps {
+  flexDirection: string;
+  flexDirectionDT: string;
+  margin: string;
+  padding: string;
+  justifyContent: string;
+  alignItems: string;
+  background: string;
+  width: string;
+}
+
+export type PropsOptional = Partial<WraperProps>;
